size_t instead of int

+++ source/src/utility/utility.cpp
@@ -176,7 +176,7 @@ void drawWithoutReplacementFisherYates(std::vector<size_t>& result, std::mt19937
 
   // Draw without replacement using Fisher Yates algorithm
   std::uniform_real_distribution<double> distribution(0.0, 1.0);
-  for (int i = 0; i < num_samples; ++i) {
+  for (size_t i = 0; i < num_samples; ++i) {
     size_t j = i + distribution(random_number_generator) * (max - skip.size() - i);
     std::swap(result[i], result[j]);
   }
